Bluetooth: Fixed Contacts Tx, Rx of unsupported files, and rejection of videos
Added code to vCardManager to handle contacts with no name and no phone number.
In the case of no name, the phone number is used as the name, and in the case
of no name and no phone number the string Unknown is used as the name.

Added code to ObexReceiverService to send authorize complete to cancel
incoming OPP transfer of unsupported file types.

Added code to ObexReceiverService to check received files to determine if
they are valid video files based on file extension