Bluetooth: Cleaned up logging tag and negative logic
As part of issue investigation, discovered debug logtag was incorrect
and file included some negative logic which was hard to understand.
Modified for cleanup purposes.

Also, changed default value returned when trying to get the SUCCESS
extra from the Bluetooth Obex intent. This defaults the functionality
to failing (do nothing) if the extra is not sent in the intent.

CRs-fixed: 211879