Bluetooth: Fix vCard RX, vCal reject, OPP authorize timeout, Add vCard for FTP
When receiving a vCard object via OPP, either client or server, with multiple
vCards embedded in the object, the data is stored incorrectly. All
information in each of the individual embedded vCards is combined and stored
in a single contact with the name of the last vCard in the object. The
receiver has been modified to pull out each embedded vCard prior to parsing
and storing. In addition, embedded vCards with the same name are combined
into a single vCard prior to parsing and storing.

Added code to reject vCals via OPP

Added timer for authorization dialog timeout to remove the dialog and signal
to the Bluetooth stack that the file was rejected

Modified code so vCards transferred via FTP are also added to contacts

CRs-fixed: 217872,217934,219205