Bluetooth: Fix to FTP and OPP concurrency
Since the methods and members of SRM are static in ObexHelper Clase, when FTP and OPP (or any two obex profiles)
are operating on SRM and non SRM modes(over L2CAP) the second session overwrites the SRM satus of first one.
To avoid the issue, members and methods of SRM are modified to operate on instance based(an instance per session).
FTP and OPP applications also modifiled accordingly.

Change-Id: Ie050821d4beaabec604ea6e44c6dc103441b4639
CRs-fixed: 298776