Bluetooth: BPP: Proper handling of broadcast receiver de-registration
If BPP transaction is interrupted either by turning off Bluetooth or going to
settings view then the sending of file is canceled. As part of the cleanup,
its now ensured that the registered broadcast receiver (for Turn off and ACL
disconnected) is unregistered prior to the BluetoothOppService destroy. This
ensures that there is no leak of IntentReceiver when BluetoothOppService
exits thereby avoiding an exception.

Change-Id: I301258fac4a617b35f177270408fa6c79c38fba8
CRs-fixed: 299740