Bluetooth: Fix for OPP crash onPut
It fixes OPP crash onPut due to the NULL pointer exception.
And changes are done to continue OPP transfer even if OPP class id is not available in the BT cache.
Because, it is expected that OPP server may not have the Device class id of client when it is acting as server
and we can not mandate the server to do device discovery before accepting the put operation.

Change-Id: I9876a69646a3c7c5ee61809d36abcb03c28c2603
CRs-fixed: 300919