Bluetooth: Fix for OPP crash onPut
It fixes OPP crash onPut due to the NULL pointer exception.
And changes are done to continue OPP transfer even if OPP class id is not available in the BT cache.
Because, it is expected that OPP server may not have the Device class id of client when it is acting as server
and we can not mandate the server to do device discovery before accepting the put operation.

CRs-fixed: 300919
(cherry picked from commit 311ace6112b103323504deb5c4db93464315241b)

Change-Id: I57832d011391511c8ae95c1f223c7f187b8be69a