Bluetooth: Fix and code optimization for FTP folder listing
1.When there is a change in string length or the number of fields returned from Java Date class(JDK) instance,
the code may leads to an unexpected behavour, since the string indexes are hard coded.
The changes are made in such a way that the string format is enforced to extract the fields from
Date class object as per the requirement. And it make sure that even changes in Java package(Date class) will
not impact the current implementation in future.
2.The private method convertMonthtoDigit() is removed as the purpose of the method is available in Java package

CRs-fixed: 310637
(cherry picked from commit 4d74bc3f9e483235fd8403899c3180ce7ef79b53)

Change-Id: If4add7d9c9feef1676f4667acfaf43f6e7b83262