Bluetooth: Reject second FTP connection with the PSM/SCN which is already in use.
FTP service creates a listener socket with PSM and SCN for L2CAP and RFCOMM respectively.
The socket handle is used to listen the connection request over RFCOMM or L2CAP from
the remote end. Once the connection is received on the socket and accepted by FTP server,
the socket is being closed to reject the second connection request on the same PSM/SCN.

Change-Id: Iae0e058b175e71b76f5735f18e8114b3bab0e6dc
CRs-fixed: 317042