Bluetooth: Clean up BPP batch and client session thread on BT off
1.When there is a BPP transfer in progress and BT is powered off in the middle,
the next transfer after BT power on is not going through.It is because, the
BPP client session thread and the batch of the previous trasfer is not
cleaned up properly. The current change handles the case.
2.Remove DOS line endings Control-M characters in all the available files

CRs-fixed: 336336
(cherry picked from commit 97c350d356cde7644c5b5392c5b757f8072542d9)

Change-Id: I2dd230ceee032353b439b9b78436a240e429feb9