Wakes device to send indications when Apps processor goes to sleep
Wakes up the device to send periodic notifications and indications
from the GATT server to the client devices when the apps
processor goes to sleep.
The device is not woken up when the notifications and indications
for all the clients are disabled or when all the client devices
receiving notifications and indications from the GATT server are
disconnected.

(cherry picked from commit 3e1c8fac054fa7a75746535e0ebc9c77cd55da77)

Change-Id: I1488331d688fd5fee58b75ca635998bc4d6d80c1