Bluetooth: Do not terminate the session when a Batch contains an empty file
If an empty file is rejected by the remote device, client session is
terminated and remaining files in the batch also marked as failure.
This change continues to send the pending files in the batch even an
empty file is rejected by the remote device. Also it provides appropriate
message to the user.

Change-Id: Id241d5aafd8120c480a7b3469131056e92c50713
CRs-fixed: 353630