Update Email account in MAS when BT turns on
Email account entry is removed when BT is turned off,
but is not restored when BT turns on again. It is updated
only when there is change in Email app or when MAP connection
is manually accepted using connection pop up. It does not
cover the case when MAP is set as trusted and BT is restarted
and connection is tried there after.

CRs-Fixed: 422255

Change-Id: I1e980c9fb3a484d256757306fb267bb692f311a0