Update Email account in MAS when BT turns on
Email account entry is removed when BT is turned off,
but is not restored when BT turns on again. It is updated
only when there is change in Email app or when MAP connection
is manually accepted using connection pop up. It does not
cover the case when MAP is set as trusted and BT is restarted
and connection is tried there after.

(cherry picked from commit 55ce29ed32fb254227a7435f2cdc4b48a771ec9a)

Change-Id: I48f2fd53a787e13c1b087a8744d8b906d5b4c96b
CRs-Fixed: 422255