Bluetooth: Do not terminate session when a Batch contains an empty file
If an empty file is rejected by the remote device, client session is
terminated and remaining files in the batch also marked as failure.
This change continues to send the pending files in the batch even an
empty file is rejected by the remote device. Also it provides appropriate
message to the user.

Change-Id: I0a48e1fb652c51b8e7eb8fa1c9a66efe99acd3d2