Bluetooth: Pbap: Acquire wakelocks only when required
This patch updates the wakelock acquiring logic in PBAP Server only when
OBEX transaction is being done by PBAP Client. Without this patch it was
observed that wakelock was acquired during obex server connection
establishement and was only released when connection was disconnected
which was resulting in extra power consumption when PBAP connection was
idle.

Change-Id: I3189036649f09c6df944d575eabb5b03863ee7d7
CRs-Fixed: 565528