Bluetooth: Guard JNI callbacks with NULL checks
With Always BLE-ON changes, profiles will be stopped before
stack is cleaned-up. Due to this JNI callbacks can arrive from the
stack even after Profile services are stopped, JNI callbacks
are guarded with NULL check before it is forwarded to JAVA layers

Below are the changes
PAN JNI is crashed due to timing, when mCallbacksObj is not allocated.
Because control_state_callback is triggered before allocating GlobalRef
for mCallbacksObj

AVRCP JNI callbacks are guarded by mCallbacksObj to avoid crashes

CRs-fixed: 655498

Change-Id: Ib04b37f6e4720d7bddc2bcde4c199cd209a79d84