Bluetooth-OPP: Fixed the BT occurs ANR when tap "try again" button
When tap the "try again" button on the send failed dialog, the
method "BluetoothOppSendFileInfo::generateFileInfo()" will be
call, if there are about 1500 contacts, execute this method
would be very time consuming, So caused the ANR happen.
Changed this method into the child thread execution, through
a Handler send a message to the main thread after this method
return, and then retry the failed transfer.

CRs-Fixed: 522213
Change-Id: I22bd04160e826789f8154bd334d948bb66ea1296