PBAP: Race-condition when requesting authorization from Settings
When MAP and PBAP requests authorization from the Settings app via
Intents, it also sets a state isWaitingForAuthorization to indicate
that it should handle incoming replies. This is however done in the
code after the intent have been sent and in some cases the reply
for the intent is returned before the state has been update.

Change-Id: I0c79bfa74465cba7af2fa50a001f487b7f74b40d
CRs-Fixed: 657103