PBAP: Update the "isWaitingAuthorization" flag properly.
The "isWaitingAuthorization" flag was updated to true in the code
after the "ACTION_CONNECTION_ACCESS_REQUEST" intent have been sent
and in some cases the reply for the intent is returned before the
state has been update. The change is done to interchange the order
of the two statements.

Change-Id: I17c5a6273521673f6412de1ecfa3045c554dbbbf
CRs-Fixed: 675107