Bluetooth: Map: Add only originator phone number in getMessage response
In a scenario where originator vcard comprises of multiple numbers,
Map server is adding all the numbers present in Originator's vcard
to bMessage while responding to getMessage request. This is creating
IOT issues with all the carkits, as carkits are not able to parse
properly this bMessage and always showing the last contact as the
message Originator. Add change to keep only the primary number with
which the message is sent, on the originator vcard.

Change-Id: I9b3368353ad57476357b4d9933227d39653f3328
CRs-fixed: 694556