Bluetooth-OPP: Handle null pointer exception during transfer failure
If OPP transfer is failed because of connection failure, OPP
will call the transfer stop to cleanup the session and connectthread.
Since OPP is failed due to TRANSPORT_ERROR that will cleanup the connect
thread, but if TRANSPORT_ERROR is received during stop that will cause
Bluetooth application crash.

CRs-Fixed: 712216
Change-Id: I595e0105bf643946fbdc370db0c2cf951250d15f