Bluetooth-OPP: Send the share status of file once transfer is complete
Share status of the file will be sent with MSG_SHARE_COMPLETE to
avoid sending the file again during batch transfer. Otherwise, if
DUT is sending batch containing small files transfer will be completed
immediately and status may not be updated to correct state. This will
lead to send same file again and cause transfer failure which is already
sent and removed from the hashmap of pending shares.

CRs-Fixed: 765769
Change-Id: I6d1b708be014fe454357c5082e3e43a640a6f7d7