Bluetooth: Killing the process in case of stop_timeout
While stopping the profiles, if stop timeout comes,
it will move back to BLE_ON states and from there
it moves to BLE_OFF state if there is no BLE app
open. This will trigger the stack shutdown while
some profiles may still be active.  We can not
move to OFF state or BLE_ON state. The only choice
is to restart the process.

CRs-Fixed: 936968
Change-Id: I5c34789b33daaf6c5e5210c8c215fa3cfbb1735d