PBAP: Remove duplicate entries fetched from database
This patch will remove the duplicate entries having same handle
and name fetched from the database sometimes. The check was
there in the stock code but that was not sufficient and was
comparing with the previous value only if these are duplicates
and not with the whole list. I have updated this check to
compare the fetched value if the exact duplicate value
(same ID and NAME) is present in the final list of contacts
which is to be returned.

CRs-Fixed: 945026
Change-Id: If32526076baed813d39ce298c307196b63bc46fd