MAP: Fix service shutdown procedure
At service stop() a signal is sent to the handler thread to shutdown any
servers and the service it self.
It however seems like the MapService instance is released for garbage
collection before the shutdown is completed. The observed behaviour is:
    * the message handler to stop handling messages
    * the object is garbage collected during shutdown of the MapServer
      instances.

The fix is to wait for shutdown to complete before returning from
stop(), and use a Map owned looper to drive the message loop.

Change-Id: Id315628668d8f6643a1a66589f9a01964e85a6a7
CRs-Fixed: 948117