HFP AG: Fix the condition to check if SCO is connected
Two headsets hs1, hs2 are connected and SCO is created with
hs2. If hs1 is disconnecting, we move to MultiHFPending state
from AudioOn state and then move back to AudioOn state. We
should check for audio state and SCO device instead of
checking if we are in AudioOn state.

CRs-Fixed: 968433
Change-Id: Ic48f4403b544ee11258d47807630a435f736d8f9