HF CLIENT: Add NULL check for call back object in JNI
When BT turned on and turned off repeatedly, if the application
thread cleans up the call back object of the app during the
stack thread upstreaming the event to the app, there can be
a crash since there is no null check while dereferencing the
call back object. Add null check before dereferencing the
call back object of the application.

CRs-Fixed: 951522
Change-Id: I9c8b94f77f8b019b13d815119aa52c080a4ec04a