BT: Killing Bluetooth process in case of disable timeout
When timeout occurs, it might happen that the disable
event does not come at all. In such case further calls
will get queued up waiting for it causing a lockup
This change will kill the app to start all the queues
so that the further commands don't get affected.

CRs-Fixed: 951917

Change-Id: I77ada2b2332aaadc224ba64ad0e3619fc94d5e72