OPP: Initialize obex socket listeners while starting Listening
This patch initializes socket listener when socket listening is to be started
and not during onCreate. Without this patch there were usecases when BT was
not bein turned off completely like in BLE on state, in these states, onDestroy
was not getting called for OppService, leading to issues during next BT on
state and would lead to OPP transfer failure.

Change-Id: Iea79c1aea53d6c4dbf1d96050902dc3eeaa5a6b1
CRs-Fixed: 978529