AVRCP(T): Fix BAD_REQUEST error sent during Get Image Request
Use Case:
1) Cotroller sends Get Image Properties request.
2) Controller sends Get ImageRequest  with max_size set in Image
   Descriptor to the native image of the requested image.

Expectation:
Request to be completed successfully, and Controller is able to receive
image within the max_size requested.

Observation:
AVRCP TG responds with BAD_REQUEST for OBEX Get Request but still sends
the image.

Root Cause:
Image data was filled in obex OutputStream, and then the check for file size
was performed, which was leading to OBEX bad request in case file size exceeded
the value requested by Controller.

Fix:
Use lower compression to compress the Image Data as 100% compression usually
exceeds the size of native image. Also add check to not add data to OBEX
OutputStream if the file size is more than the requested limit.

Change-Id: I370a1878d27f0441002816fc26d226b037205f8b
CRs-Fixed: 993842