Bluetooth: carkit doesn't show music name information
The media number and totoal media number of some music
equal to zero, When DUT replied the get media information
fit the media number field as "0" string instead of an
empty string, some kind carkit cannot handle this and
display music name correctlly. In this fix, DUT reply
an empty string when the number value is 0.

Change-Id: If7810c103ebaca8de8706844238b7c20febe9411
CRs-Fixed: 1104929