Resume A2DP playback if suspended from HFP during VR
When VR is started and stopped very fast, there is
a chance that A2DP is suspended for SCO for VR.
HeadStateMachine is still in Connected state. When
VR is stopped, A2DP is not resumed since SCO gets
connected/disconnected in Connected state.

Wait for A2DP suspended broadcast to create SCO
during VR.

CRs-Fixed: 2005951
Change-Id: Iee6452570d99f3638b4c9e0e2f647c7b8746eeec