PBAP: Fix to fetch phonebook vcards when startpoint is contact size - 1.
Usecase:
Send Get request from PBAP client to fetch vcards with startoffset
and MaxListCount selected in such a way that startPoint matches
[n(contacts) -1] for any number of contacts.
Ex: startOffset=0, MaxListCount=2.

Issue:
Pull Phonebook vcards for above condition was failing.

Root Cause:
startPoint was checked with ContactId of contact which resulted in
contacts not getting fetched in the given range.

Fix:
startPoint is compared with index of contact instead of ContactId.

Change-Id: I435313eee00e3dae135a6021b45b47a37d80d0d8
CRs-Fixed: 2011445