AVRCP: Fix for remote sending pause during soft handoff
If the AVRCP connection is delayed for the second connected device
and the second connected device is a playing device then
non-playing device's play status is updated as playing. If the user
tries to trigger soft handoff, non-playing device always send
pause instead of play as its play status is updated as playing,
resulting in soft handoff not working. To fix this issue, when
the avrcp connection comes up for the playing device, update the
correct play status to non-playing device.

Change-Id: I86d1bb4d643d225deb573e55624c7f9ed51b3330